% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deemdat.R
\name{deemdat}
\alias{deemdat}
\title{Denoising Ensemble Empirical Mode Decomposition and Adaptive Thresholding}
\usage{
deemdat(x, sigma, m1, m2, ...)
}
\arguments{
\item{x}{time serie (ts)}

\item{sigma}{main parameter to be set}

\item{m1}{the first (m1 - 1) IMFs are removed}

\item{m2}{applied between the m1-th and (M - m2)th IMFs}

\item{...}{additional arguments to be passed to \code{\link[Rlibeemd]{eemd}} function}
}
\description{
random noise removal by band-pass filtering in 
the time domain based on ensemble empirical mode 
decomposition (EEMD) combined with adaptive thresholding.
}
\details{
thresholding is only applied between the m1-th 
and (M - m2)th IMFs, and the first (m1 - 1) IMFs are removed,
where IMFk is the kth IMF andMis the total number of IMFs of 
the input signal. If m2 is set to 0, we apply the thresholding
from the m1th IMF to the last IMF.
}
\references{
HAN, J.; VAN DER BAAN, M. Microseismic and seismic denoising via ensemble empirical mode decomposition and adaptive thresholding. Geophysics, v. 80, n. 6, p. KS69-KS80, 2015.
}

