% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deemdat.R
\name{iht}
\alias{iht}
\title{interval hard thresholding}
\usage{
iht(h, Th)
}
\arguments{
\item{h}{input signal}

\item{Th}{universal threshold}
}
\value{
hz thresholded signal
}
\description{
The idea of IMF interval thresholding is maintaining the
 whole interval between two zero crossings in each IMF, 
 when the absolute value of local extrema in this interval 
 is larger than the threshold.
}
\references{
HAN, J.; VAN DER BAAN, M. Microseismic and seismic denoising via ensemble empirical mode decomposition and adaptive thresholding. Geophysics, v. 80, n. 6, p. KS69-KS80, 2015.
}

